fix: don't authenticate OPTIONS calls
This fixes issues when using the BrowsableAPIRenderer from DRF that
would result in an action of 'None' and raise an error unless you
specifically add a mapping for 'None'.

This was only a problem for single object permissions - listing already
had a special case for None. I have changed the special case to be for
the specific 'OPTIONS' method instead to be more explicit.